requirements.txt
streamlit
python-dotenv
langchain
langchain-community
langchain-groq
fpdf